backend/requirements.txt
pandas==2.2.1
Pillow==10.2.0
protobuf==4.25.3
google-cloud-vision==3.7.2
google-cloud-texttospeech==2.16.3
boto3==1.34.61